Reuse shared language modal on login
